+++ Changelog.rst
@@ -3,9 +3,20 @@ Changelog
 
 Summary of all changes made since the first stable release
 
+0.5.0 (01-28-2025)
+------------------
+* DEP: Removed deprecated functions that depend on ssj_auroral_boundary package
+* DEP: Removed deprecated methods, attributes, and kwargs in VectorData
+* DOC: Added data reference for SuperDARN vorticity data
+* BUG: Fixed `load_vorticity_ascii_data` to correctly cycle through files
+* BUG: Fixed EAB index pairing issue in `DualBoundary` class
+* BUG: Fixed `zenodo-get` import in pyproject.toml
+* TST: Updated GitHub Actions yamls
+* MAINT: Cycled Python and pysat version support
+
 0.4.0 (06-07-2024)
 ------------------
-* DEP: Deprecated functions that depend on ssj_auroral_boundary_package
+* DEP: Deprecated functions that depend on ssj_auroral_boundary package
 * DEP: Removed classes and kwargs deprecated in v0.3.0
 * MAINT: Cycled supported Python versions
 * MAINT: Added a pyproject.toml file and removed setup.py